# api/openapi.yaml
openapi: 3.0.3
info:
  title: Trellis API
  description: |
    Trellis is a development environment control plane that unifies local development,
    parallel branches, remote terminals, and live production logs.

    ## Versioning

    The API uses URL path versioning (`/api/v1/...`) for major versions. Within a major
    version, Stripe-style date-based versioning is supported via the `Trellis-Version`
    HTTP header (e.g., `Trellis-Version: 2026-01-17`).

    By default, the latest behavior is used. Pin to a specific date version for stability
    when breaking changes are introduced within a major version. The version date represents
    when a particular API behavior was current.

    ## Go Client Library

    An official Go client library is available at `github.com/wingedpig/trellis/pkg/client`.
    It provides typed access to all endpoints and supports API versioning. See the
    [documentation](/docs/reference/client/) for usage examples.
  version: 1.0.0
  contact:
    name: Trellis
    url: https://trellis.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

# Trellis is a local development tool with no authentication
security: []
...
paths:
...
  /services/{name}:
    get:
      tags: [Services]
      summary: Get service details
      operationId: getService
      parameters:
        - $ref: '#/components/parameters/ServiceName'
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ServiceInfo'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worktrees:
    get:
      tags: [Worktrees]
      summary: List all worktrees
      operationId: listWorktrees
      parameters:
        - $ref: '#/components/parameters/TrellisVersion'
      responses:
        '200':
          description: List of worktrees
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      worktrees:
                        type: array
                        items:
                          $ref: '#/components/schemas/Worktree'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
    post:
      tags: [Worktrees]
      summary: Create a new worktree
      operationId: createWorktree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_name]
              properties:
                branch_name:
                  type: string
                  description: Branch name to create worktree from
                switch_to:
                  type: boolean
                  default: false
                  description: Whether to activate the new worktree
      responses:
        '201':
          description: Worktree created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      worktree:
                        $ref: '#/components/schemas/Worktree'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /worktrees/{name}:
    get:
      tags: [Worktrees]
      summary: Get worktree details
      operationId: getWorktree
      parameters:
        - $ref: '#/components/parameters/WorktreeName'
      responses:
        '200':
          description: Worktree details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Worktree'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Worktrees]
      summary: Remove a worktree
      operationId: removeWorktree
      parameters:
        - $ref: '#/components/parameters/WorktreeName'
        - name: delete_branch
          in: query
          description: Whether to delete the associated branch
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Worktree removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      removed:
                        type: string
                      branch_deleted:
                        type: boolean
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    TrellisVersion:
      name: Trellis-Version
      in: header
      required: false
      description: |
        API version date (e.g., 2026-01-17). This header can be sent with any API request.
        When omitted, the latest API behavior is used. Pin to a specific date for stability
        across API updates. The Go client library sets this automatically via WithVersion().
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-01-17'
    ServiceName:
      name: name
      in: path
      required: true
      description: Service name
      schema:
        type: string
    WorktreeName:
      name: name
      in: path
      required: true
      description: Worktree name
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ServiceInfo:
      type: object
      properties:
        Name:
          type: string
        Status:
          $ref: '#/components/schemas/ServiceStatus'
        Enabled:
          type: boolean
        ParserType:
          type: string
        Layout:
          type: array
          items:
            type: object
        TimestampField:
          type: string
        LevelField:
          type: string
        MessageField:
          type: string
    ServiceStatus:
      type: object
      properties:
        State:
          type: string
          enum: [stopped, starting, running, stopping, crashed]
        PID:
          type: integer
        ExitCode:
          type: integer
        StartedAt:
          type: string
          format: date-time
        StoppedAt:
          type: string
          format: date-time
        RestartCount:
          type: integer
        Error:
          type: string
    Worktree:
      type: object
      properties:
        Path:
          type: string
        Branch:
          type: string
        Commit:
          type: string
        Detached:
          type: boolean
        IsBare:
          type: boolean
        Dirty:
          type: boolean
        Ahead:
          type: integer
        Behind:
          type: integer
        Active:
          type: boolean
...